# api/openapi.yaml
openapi: 3.0.3
info:
  title: Arc Issue Tracker API
  description: |
    Central issue tracking server for AI-assisted coding workflows.
    Arc provides project-scoped issue management with dependency tracking,
    labels, comments, and audit trails.
  version: 1.0.0
  contact:
    name: SentioLabs
    url: https://github.com/sentiolabs/arc
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      responses:
        "204":
          description: Project deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/stats:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project statistics
      responses:
        "200":
          description: Project statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statistics"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Issues
  # ====================
  /projects/{projectId}/issues:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: listIssues
      tags: [issues]
      summary: List issues in project
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            $ref: "#/components/schemas/Status"
        - name: type
          in: query
          description: Filter by issue type
          schema:
            $ref: "#/components/schemas/IssueType"
        - name: priority
          in: query
          description: Filter by priority (0-4)
          schema:
            type: integer
            minimum: 0
            maximum: 4
        - name: assignee
          in: query
          description: Filter by assignee
          schema:
            type: string
        - name: parent_id
          in: query
          required: false
          description: Filter by parent issue ID (returns children via parent-child dependency)
          schema:
            type: string
        - name: q
          in: query
          description: Full-text search query
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of issues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIssues"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createIssue
      tags: [issues]
      summary: Create a new issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueRequest"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    get:
      operationId: getIssue
      tags: [issues]
      summary: Get issue by ID
      parameters:
        - name: details
          in: query
          description: Include full details (dependencies, comments, labels)
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Issue details
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Issue"
                  - $ref: "#/components/schemas/IssueDetails"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateIssue
      tags: [issues]
      summary: Update issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueRequest"
      responses:
        "200":
          description: Issue updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete issue
      responses:
        "204":
          description: Issue deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/close:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    post:
      operationId: closeIssue
      tags: [issues]
      summary: Close an issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseIssueRequest"
      responses:
        "200":
          description: Issue closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/reopen:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    post:
      operationId: reopenIssue
      tags: [issues]
      summary: Reopen a closed issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      responses:
        "200":
          description: Issue reopened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Ready & Blocked
  # ====================
...
  /projects/{projectId}/blocked:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: getBlockedIssues
      tags: [ready]
      summary: Get blocked issues
      parameters:
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        "200":
          description: List of blocked issues with blocking info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedIssue"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Team Context
  # ====================
  /projects/{projectId}/team-context:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: getTeamContext
      tags: [teams]
      summary: Get issues grouped by teammate role labels
      description: |
        Returns issues grouped by their `teammate:*` labels. When an epic_id is
        provided, only children of that epic (via parent-child dependency) are
        included. Without epic_id, all non-closed issues with teammate labels
        in the project are returned.
      parameters:
        - name: epic_id
          in: query
          description: Optional epic ID to scope to children of a specific epic
          schema:
            type: string
      responses:
        "200":
          description: Team context with issues grouped by role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Dependencies
  # ====================
  /projects/{projectId}/issues/{issueId}/deps:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    get:
      operationId: getDependencies
      tags: [dependencies]
      summary: Get issue dependencies and dependents
      responses:
        "200":
          description: Dependencies and dependents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DependencyGraph"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: addDependency
      tags: [dependencies]
      summary: Add a dependency to an issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDependencyRequest"
      responses:
        "201":
          description: Dependency added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dependency"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/deps/{dependsOnId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"
      - name: dependsOnId
        in: path
        required: true
        description: ID of the dependency to remove
        schema:
          type: string

    delete:
      operationId: removeDependency
      tags: [dependencies]
      summary: Remove a dependency from an issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      responses:
        "204":
          description: Dependency removed
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Labels (global)
  # ====================
  /labels:
    get:
      operationId: listLabels
      tags: [labels]
      summary: List all global labels
      responses:
        "200":
          description: List of labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createLabel
      tags: [labels]
      summary: Create a new label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelRequest"
      responses:
        "201":
          description: Label created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /labels/{labelName}:
    parameters:
      - name: labelName
        in: path
        required: true
        description: Label name
        schema:
          type: string

    put:
      operationId: updateLabel
      tags: [labels]
      summary: Update a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabelRequest"
      responses:
        "200":
          description: Label updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteLabel
      tags: [labels]
      summary: Delete a label
      responses:
        "204":
          description: Label deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Issue-Label Associations (project-scoped)
  # ====================
  /projects/{projectId}/issues/{issueId}/labels:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    post:
      operationId: addLabelToIssue
      tags: [labels]
      summary: Add a label to an issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLabelToIssueRequest"
      responses:
        "204":
          description: Label added to issue
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/labels/{labelName}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"
      - name: labelName
        in: path
        required: true
        description: Label name to remove
        schema:
          type: string

    delete:
      operationId: removeLabelFromIssue
      tags: [labels]
      summary: Remove a label from an issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      responses:
        "204":
          description: Label removed from issue
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Comments & Events
  # ====================
  /projects/{projectId}/issues/{issueId}/comments:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    get:
      operationId: getComments
      tags: [comments]
      summary: Get comments for an issue
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment to an issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/comments/{commentId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"
      - name: commentId
        in: path
        required: true
        description: Comment ID
        schema:
          type: integer
          format: int64

    put:
      operationId: updateComment
      tags: [comments]
      summary: Update a comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "204":
          description: Comment updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      responses:
        "204":
          description: Comment deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/events:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    get:
      operationId: getEvents
      tags: [comments]
      summary: Get audit events for an issue
      parameters:
        - name: limit
          in: query
          description: Maximum events to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
    IssueId:
      name: issueId
      in: path
      required: true
      description: Issue ID
      schema:
        type: string
    ActorHeader:
      name: X-Actor
      in: header
      description: User performing the action (defaults to "anonymous")
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Status:
      type: string
      enum:
        - open
        - in_progress
        - blocked
        - deferred
        - closed
    IssueType:
      type: string
      enum:
        - bug
        - feature
        - task
        - epic
        - chore
...
    Project:
      type: object
      required:
        - id
        - name
        - prefix
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique project ID (e.g., "proj-a1b2")
        name:
          type: string
          description: Display name
          maxLength: 100
        path:
          type: string
          description: Optional associated directory path
        description:
          type: string
          description: Project description
        prefix:
          type: string
          description: Issue ID prefix (e.g., "bd")
          maxLength: 10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required:
        - name
        - prefix
      properties:
        name:
          type: string
          description: Display name
          maxLength: 100
        path:
          type: string
          description: Optional associated directory path
        description:
          type: string
          description: Project description
        prefix:
          type: string
          description: Issue ID prefix
          maxLength: 10
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          description: New display name
          maxLength: 100
        path:
          type: string
          description: New associated directory path
        description:
          type: string
          description: New description
    Statistics:
      type: object
      required:
        - project_id
        - total_issues
        - open_issues
        - in_progress_issues
        - closed_issues
        - blocked_issues
        - deferred_issues
        - ready_issues
      properties:
        project_id:
          type: string
        total_issues:
          type: integer
        open_issues:
          type: integer
        in_progress_issues:
          type: integer
        closed_issues:
          type: integer
        blocked_issues:
          type: integer
        deferred_issues:
          type: integer
        ready_issues:
          type: integer
        avg_lead_time_hours:
          type: number
          format: double

    # ====================
    # Issue Schemas
    # ====================
    Issue:
      type: object
      required:
        - id
        - project_id
        - title
        - status
        - priority
        - issue_type
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique issue ID
        project_id:
          type: string
        title:
          type: string
          maxLength: 500
        description:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        priority:
          type: integer
          minimum: 0
          maximum: 4
          description: "0 (critical) - 4 (backlog)"
        rank:
          type: integer
          description: "0 = unranked (sorts last), 1+ = lower rank = work on first"
        issue_type:
          $ref: "#/components/schemas/IssueType"
        assignee:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        close_reason:
          type: string
        external_ref:
          type: string
          description: External reference (e.g., "gh-9", "jira-ABC")
        labels:
          type: array
          items:
            type: string
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/Dependency"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    IssueDetails:
      allOf:
        - $ref: "#/components/schemas/Issue"
        - type: object
          properties:
            dependents:
              type: array
              items:
                $ref: "#/components/schemas/Dependency"
...
    PaginatedIssues:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ====================
    # Team Context Schemas
    # ====================
...
    CreateIssueRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        priority:
          type: integer
          minimum: 0
          maximum: 4
          default: 2
        issue_type:
          $ref: "#/components/schemas/IssueType"
        assignee:
          type: string
        external_ref:
          type: string
    UpdateIssueRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        priority:
          type: integer
          minimum: 0
          maximum: 4
        issue_type:
          $ref: "#/components/schemas/IssueType"
        assignee:
          type: string
        external_ref:
          type: string
    CloseIssueRequest:
      type: object
      properties:
        reason:
          type: string
          description: Reason for closing

    # ====================
    # Dependency Schemas
    # ====================
...